+++ tip/src/bin/tsfreg.rs
@@ -20,7 +20,8 @@ const ILOT_INSTALL: u32 = 0x00000000;
 // const ILOT_UNINSTALL: u32 = 0x00000001;
 
 const CHEWING_TSF_CLSID: GUID = GUID::from_u128(0x13F2EF08_575C_4D8C_88E0_F67BB8052B84);
-const CHEWING_PROFILE_GUID: GUID = GUID::from_u128(0xCE45F71D_CE79_41D1_967D_640B65A380E3);
+const CHEWING_ZH_TW_PROFILE_GUID: GUID = GUID::from_u128(0xCE45F71D_CE79_41D1_967D_640B65A380E3);
+const CHEWING_ZH_CN_PROFILE_GUID: GUID = GUID::from_u128(0x9063E7CC_225E_4CEA_B7A1_2B0FB12A75CF);
 const CHEWING_TIP_DESC: PCWSTR =
     w!("0x0404:{13F2EF08-575C-4D8C-88E0-F67BB8052B84}{CE45F71D-CE79-41D1-967D-640B65A380E3}");
 
@@ -38,18 +39,35 @@ fn register(icon_path: String) -> Result<()> {
         let input_processor_profile_mgr: ITfInputProcessorProfileMgr =
             CoCreateInstance(&CLSID_TF_InputProcessorProfiles, None, CLSCTX_INPROC_SERVER)?;
 
-        let mut lcid = LocaleNameToLCID(w!("zh-Hant-TW"), 0);
-        if lcid == 0 {
-            lcid = LocaleNameToLCID(w!("zh-TW"), 0);
-        }
-
         let pw_icon_path = icon_path.encode_utf16().collect::<Vec<_>>();
 
+        // Register for zh_TW
+        let mut lcid = LocaleNameToLCID(w!("zh-TW"), 0);
+        if matches!(lcid, 0 | 0x0C00 | 0x1000) {
+            lcid = 0x404;
+        }
         input_processor_profile_mgr.RegisterProfile(
             &CHEWING_TSF_CLSID,
             lcid as u16,
-            &CHEWING_PROFILE_GUID,
+            &CHEWING_ZH_TW_PROFILE_GUID,
             w!("新酷音輸入法").as_wide(),
+            &pw_icon_path,
+            0,
+            HKL::default(),
+            0,
+            false,
+            0,
+        )?;
+        // Register for zh_CN
+        let mut lcid = LocaleNameToLCID(w!("zh-CN"), 0);
+        if matches!(lcid, 0 | 0x0C00 | 0x1000) {
+            lcid = 0x804;
+        }
+        input_processor_profile_mgr.RegisterProfile(
+            &CHEWING_TSF_CLSID,
+            lcid as u16,
+            &CHEWING_ZH_CN_PROFILE_GUID,
+            w!("新酷音输入法").as_wide(),
             &pw_icon_path,
             0,
             HKL::default(),
@@ -92,15 +110,26 @@ fn unregister() -> Result<()> {
             }
         }
 
-        let mut lcid = LocaleNameToLCID(w!("zh-Hant-TW"), 0);
-        if lcid == 0 {
-            lcid = LocaleNameToLCID(w!("zh-TW"), 0);
+        // Unregister zh_TW profile
+        let mut lcid = LocaleNameToLCID(w!("zh-TW"), 0);
+        if matches!(lcid, 0 | 0x0C00 | 0x1000) {
+            lcid = 0x404;
         }
-
         input_processor_profile_mgr.UnregisterProfile(
             &CHEWING_TSF_CLSID,
             lcid as u16,
-            &CHEWING_PROFILE_GUID,
+            &CHEWING_ZH_TW_PROFILE_GUID,
+            0,
+        )?;
+        // Unregister zh_CN profile
+        let mut lcid = LocaleNameToLCID(w!("zh-CN"), 0);
+        if matches!(lcid, 0 | 0x0C00 | 0x1000) {
+            lcid = 0x804;
+        }
+        input_processor_profile_mgr.UnregisterProfile(
+            &CHEWING_TSF_CLSID,
+            lcid as u16,
+            &CHEWING_ZH_CN_PROFILE_GUID,
             0,
         )?;
     }