+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 - tip: re-enable support for completely disable input processing via Ctrl+Space.
 - tip: initial support for using Ctrl+Delete to unlearn phrases.
 - tip: enable chewing.auto_snapshot_selections by default.
+- tip: support zh-CN language profile.
 - prefs: preferences now only show fonts that can render Chinese characters.
 - prefs: new advanced option to lock Chinese or English mode on CapsLock.
 - prefs: new advanced option to customize candidate list border color.